Refactor of Xtext-independent classes in com.avaloq.tools.ddk.xtext
In addition, unused classes in com.avaloq.tools.ddk.xtext.util have been
removed
