+++ com.avaloq.tools.ddk.check.runtime.core/META-INF/MANIFEST.MF
@@ -10,7 +10,8 @@ Require-Bundle: org.eclipse.core.resources,
  org.eclipse.xtext,
  org.eclipse.xtext.xbase.lib,
  org.apache.commons.lang,
- com.avaloq.tools.ddk.xtext
+ com.avaloq.tools.ddk.xtext,
+ com.avaloq.tools.ddk
 Export-Package: com.avaloq.tools.ddk.check.runtime,
  com.avaloq.tools.ddk.check.runtime.configuration,
  com.avaloq.tools.ddk.check.runtime.context,
