+++ com.avaloq.tools.ddk/META-INF/MANIFEST.MF
@@ -10,8 +10,12 @@ Require-Bundle: com.google.guava,
  org.apache.commons.lang,
  org.eclipse.jdt.annotation,
  org.eclipse.core.resources,
- org.eclipse.core.runtime
-Export-Package: com.avaloq.tools.ddk.caching
+ org.eclipse.core.runtime,
+ com.google.inject
+Export-Package: com.avaloq.tools.ddk.annotations,
+ com.avaloq.tools.ddk.caching,
+ com.avaloq.tools.ddk.util,
+ com.avaloq.tools.ddk.tracing
 Import-Package: org.apache.logging.log4j,org.apache.log4j
 Automatic-Module-Name: com.avaloq.tools.ddk
