+++ com.avaloq.tools.ddk/src/com/avaloq/tools/ddk/tracing/FullGarbageCollectionEvent.java
@@ -9,7 +9,7 @@
  *     Avaloq Group AG - initial API and implementation
  *******************************************************************************/
 
-package com.avaloq.tools.ddk.xtext.tracing;
+package com.avaloq.tools.ddk.tracing;
 
 import java.util.Arrays;
