+++ com.avaloq.tools.ddk/src/com/avaloq/tools/ddk/tracing/TraceConfiguration.java
@@ -9,7 +9,7 @@
  *     Avaloq Group AG - initial API and implementation
  *******************************************************************************/
 
-package com.avaloq.tools.ddk.xtext.tracing;
+package com.avaloq.tools.ddk.tracing;
 
 import java.util.Map;
